chore: delete config.js & add to .gitignore
I have regenerated the API keys for the project. The web key is
currently still visible in the `frontend` project but I will change to
using the reserved URLs to avoid the need for the config file, and
in the meantime this prevents the correct key from being added to
github.